RUM-7880: Update TestUtilities scheme
