--- DatadogWebViewTracking/Tests/Mocks.swift
@@ -1,46 +1,0 @@
-/*
- * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
- * This product includes software developed at Datadog (https://www.datadoghq.com/).
- * Copyright 2019-Present Datadog, Inc.
- */
-
-import Foundation
-#if !os(tvOS)
-import WebKit
-import TestUtilities
-
-@testable import DatadogWebViewTracking
-
-final class DDUserContentController: WKUserContentController {
-    typealias NameHandlerPair = (name: String, handler: WKScriptMessageHandler)
-    private(set) var messageHandlers = [NameHandlerPair]()
-
-    override func add(_ scriptMessageHandler: WKScriptMessageHandler, name: String) {
-        messageHandlers.append((name: name, handler: scriptMessageHandler))
-    }
-
-    override func removeScriptMessageHandler(forName name: String) {
-        messageHandlers = messageHandlers.filter {
-            return $0.name != name
-        }
-    }
-}
-
-final class MockMessageHandler: NSObject, WKScriptMessageHandler {
-    func userContentController(_ userContentController: WKUserContentController, didReceive message: WKScriptMessage) { }
-}
-
-final class MockScriptMessage: WKScriptMessage {
-    private let _body: Any
-    private weak var _webView: WKWebView?
-
-    init(body: Any, webView: WKWebView? = nil) {
-        _body = body
-        _webView = webView
-    }
-
-    override var body: Any { _body }
-    override weak var webView: WKWebView? { _webView }
-}
-
-#endif
